:recycle:  refactor printer code to use isDeprecated
