+++ packages/printer-legacy/src/const/options.js
@@ -1,0 +1,20 @@
+const OPTION_DEPRECATED = {
+  DEFAULT: "default",
+  GROUP: "group",
+};
+
+const DEFAULT_OPTIONS = {
+  schema: undefined,
+  basePath: "/",
+  groups: {},
+  parentTypePrefix: true,
+  relatedTypeSection: true,
+  typeBadges: true,
+  skipDocDirective: undefined,
+  printDeprecated: OPTION_DEPRECATED.DEFAULT,
+};
+
+module.exports = {
+  OPTION_DEPRECATED,
+  DEFAULT_OPTIONS,
+};
