+++ packages/printer-legacy/tests/unit/graphql/input.test.js
@@ -3,7 +3,8 @@ const {
   GraphQLString,
   GraphQLInputObjectType,
 } = require("graphql");
-const { DEFAULT_OPTIONS } = require("../../../src/printer");
+
+const { DEFAULT_OPTIONS } = require("../../../src/const/options");
 
 const {
   printCodeInput,
