+++ packages/printer-legacy/tests/unit/graphql/interface.test.js
@@ -3,7 +3,8 @@ const {
   GraphQLString,
   GraphQLInterfaceType,
 } = require("graphql");
-const { DEFAULT_OPTIONS } = require("../../../src/printer");
+
+const { DEFAULT_OPTIONS } = require("../../../src/const/options");
 
 const {
   printCodeInterface,
