+++ packages/printer-legacy/tests/unit/graphql/operation.test.js
@@ -2,7 +2,7 @@ const { GraphQLID, GraphQLObjectType, GraphQLString } = require("graphql");
 
 const Utils = require("@graphql-markdown/utils");
 
-const { DEFAULT_OPTIONS } = require("../../../src/printer");
+const { DEFAULT_OPTIONS } = require("../../../src/const/options");
 const { printOperationMetadata } = require("../../../src/graphql/operation");
 
 describe("operation", () => {
